Reorganizing file structure.  Not done.
