+++ lib/ansible/inventory/__init__.py
@@ -20,13 +20,13 @@
 import fnmatch
 import os
 
-import constants as C
 import subprocess
-from ansible.inventory_parser import InventoryParser
-from ansible.inventory_parser_yaml import InventoryParserYaml
-from ansible.inventory_script import InventoryScript
-from ansible.group import Group
-from ansible.host import Host
+import ansible.constants as C
+from ansible.inventory.ini import InventoryParser
+from ansible.inventory.yaml import InventoryParserYaml
+from ansible.inventory.script import InventoryScript
+from ansible.inventory.group import Group
+from ansible.inventory.host import Host
 from ansible import errors
 from ansible import utils
