+++ lib/ansible/inventory/group.py
