+++ lib/ansible/inventory/host.py
