+++ lib/ansible/inventory/script.py
@@ -19,10 +19,9 @@
 
 import os
 import subprocess
-import constants as C
-import os
-from ansible.host import Host
-from ansible.group import Group
+import ansible.constants as C
+from ansible.inventory.host import Host
+from ansible.inventory.group import Group
 from ansible import errors
 from ansible import utils
