+++ lib/ansible/runner/connection.py
